requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
firebase-admin>=6.0.0
google-cloud-firestore>=2.0.0
requests>=2.25.0
python-dotenv>=1.0.0
Pillow>=9.0.0
numpy>=1.24.0
opencv-python>=4.8.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
            